update 0427 更新v8的config和脚本，并添加基于stage3v11的online_v1 - online_v3

+++ xtuner/v1/engine/vision_compose_train_engine.py
@@ -257,8 +257,8 @@ class VisionComposeTrainEngine(TrainEngine):
                         step_consumed_img_tokens /= seq_ctx.sequence_parallel_mesh.size()
 
                 num_tokens = seq_ctx.cu_seq_lens_k[1:] - seq_ctx.cu_seq_lens_k[:-1]
-                efficient_forward_tokens += (num_tokens**2).sum()
-                total_forward_tokens += (num_tokens.sum()) ** 2
+                efficient_forward_tokens += (num_tokens.long()**2).sum()
+                total_forward_tokens += (num_tokens.sum().long()) ** 2
             # logger.info(f"开始forward{i}个micro batch!")
             # todo: support intra_layer_micro_batch
             output = self.model(seq_ctx=seq_ctx_list[0], loss_ctx=loss_ctx_list[0])
